+++ website/docusaurus.config.js
@@ -16,6 +16,7 @@ const packages = [
     'memory-storage',
     'utils',
     'types',
+    'impit-client',
 ];
 const packagesOrder = [
     '@crawlee/core',
@@ -31,6 +32,7 @@ const packagesOrder = [
     '@crawlee/browser-pool',
     '@crawlee/utils',
     '@crawlee/types',
+    '@crawlee/impit-client',
 ];
 
 /** @type {Partial<import('@docusaurus/types').DocusaurusConfig>} */
@@ -53,10 +55,14 @@ module.exports = {
     },
     onBrokenLinks: 'throw',
     markdown: {
+        mermaid: true,
         hooks: {
             onBrokenMarkdownLinks: 'throw',
         },
     },
+    themes: [
+        '@docusaurus/theme-mermaid',
+    ],
     future: {
         experimental_faster: {
             // ssgWorkerThreads: true,
@@ -266,8 +272,8 @@ module.exports = {
             },
         },
         announcementBar: {
-            id: `crawlee-for-python-v1`,
-            content: `<b><a href="https://github.com/apify/crawlee-python">Crawlee for Python v1.0</a></b> is LIVE. 🎉️ <b><a href="https://www.producthunt.com/posts/crawlee-for-python-v1-0">Support us on Product Hunt!</a> 🥳`,
+            id: `apify-1m-challenge`,
+            content: `<b><a href="https://apify.com/challenge">Apify $1M Challenge 💰</a></b> Earn and win building with Crawlee!`,
         },
         navbar: {
             hideOnScroll: true,